+++ CHANGELOG.md
@@ -35,6 +35,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.12.2] 2019-12-11
 
 ### Added
+- Added `intersection_segment_polyline` to `compas.geometry.intersections`
+- Added `intersection_segment_polyline_xy` to `compas.geometry.intersections`
 
 ### Changed
 - Reworked docstrings of methods in `compas.geometry.queries`